fix `expires_in` type on the native side

+++ android/src/main/kotlin/com/sap/gigya_flutter_plugin/GigyaSDKWrapper.kt
@@ -434,7 +434,7 @@ class GigyaSDKWrapper<T : GigyaAccount>(application: Application, accountObj: Cl
             val sessionInfo = SessionInfo(
                 argumentMap["sessionSecret"] as String,
                 argumentMap["sessionToken"] as String,
-                (argumentMap["expires_in"] as Double).toLong()
+                argumentMap["expires_in"] as Long,
             )
             sdk.setSession(sessionInfo)
             channelResult.success(null)
